third_party/blink/web_tests/flag-specific/highdpi/fast/scroll-behavior/smooth-scroll/clamp-scroll-offset-if-size-changes-during-animation-expected.txt

This is a testharness.js-based test.
FAIL Clamps the scroll offset if the scroller's size is changed during animation. assert_equals: expected 1015 but got 1014.6666870117188
Harness: the test ran to completion.
